Add files via upload
So on the way to optimization, queue is implemented as circular queue. Quick implementation of Circular queue is here now! Here we are managing the FRONT & REAR ends to make queue circular. The objective to use it is to efficiently manage data in situations where the queue is continuously used — that is, items are repeatedly inserted and deleted — without wasting memory.
#Why We Need It
In a regular (linear) queue,
When elements are dequeued (removed), the front moves forward.
But even if there’s free space at the beginning, new elements can’t be added once the rear reaches the end.

💡 This causes wasted space.